added the actions into a returning function to keep the provider function clean
